Refactor nexus.py
To separate out functions and classes not related to reading NeXus files
